Move ByteArrayToBase64TypeAdapter to new package
Minor follow up to #3285 to move the base 64 serializer to a json
subpackage under util
